refactor: folder struct for ProjectDash -- one happy family
